Add utsarga-apavada inference
This is a monolithic commit that makes a variety of changes to the
code base. Had I remembered to commit an old file along the way,
I would have been able to separate this commit into a few more
reasonable chunks. But as it is the commit is over-ambitious.

First, the commit adds basic utsarga-apavada inference to the system,
which cleans up some code in apply_next_rule.

Next, the commit cleans up a lot of code in the code base, with an eye
to making everything more regular and predictable. Among other things,
this commit does the following:
- remove StateRule rules entirely
- add cleaner support for optional rules
- start sorting rules according to adhyaya
- remove some redundant code and old comments
- regularize filter matching for substitution rules

The result is truer to the spirit of the Ashtadhyayi and is also a
fair sight more readable.

The 'it' rules aren't fully working, but they're good enough for
most cases.

About 50% of the tests pass.