Add more advanced utsarga-apavada detection
Originally I set out to solve just the issue between 6.4.77 and 6.4.88,
but this issue touched so many parts of the system that it was hard to
change just one part of it. Along the way, this commit adds some new
tests for filters and U-A inference.

About 80% of the tests pass.