Add weak support for 'lrt'
For a change of pace. Most of the tests fail due to issues with 'it',
but some fail due to the way options are handled by the system. More
specifically, the system has assumed that if an option is declined,
then it cannot block an utsarga. But if the option and utsarga have
the same operation, then the option *must* block the utsarga.